Add attribute utils to new file
