+++ sdk/include/opentelemetry/sdk/trace/span_data.h
@@ -6,6 +6,7 @@
 #include "opentelemetry/common/attribute_value.h"
 #include "opentelemetry/core/timestamp.h"
 #include "opentelemetry/nostd/string_view.h"
+#include "opentelemetry/sdk/trace/attribute_utils.h"
 #include "opentelemetry/sdk/trace/recordable.h"
 #include "opentelemetry/trace/canonical_code.h"
 #include "opentelemetry/trace/span_id.h"
@@ -16,110 +17,6 @@ namespace sdk
 {
 namespace trace
 {
-/**
- * A counterpart to AttributeValue that makes sure a value is owned. This
- * replaces all non-owning references with owned copies.
- */
-using SpanDataAttributeValue = nostd::variant<bool,
-                                              int64_t,
-                                              uint64_t,
-                                              double,
-                                              std::string,
-                                              std::vector<bool>,
-                                              std::vector<int64_t>,
-                                              std::vector<uint64_t>,
-                                              std::vector<double>,
-                                              std::vector<std::string>>;
-
-/**
- * Creates an owned copy (SpanDataAttributeValue) of a non-owning AttributeValue.
- */
-struct AttributeConverter
-{
-  SpanDataAttributeValue operator()(bool v) { return SpanDataAttributeValue(v); }
-  SpanDataAttributeValue operator()(int v)
-  {
-    return SpanDataAttributeValue(static_cast<int64_t>(v));
-  }
-  SpanDataAttributeValue operator()(int64_t v) { return SpanDataAttributeValue(v); }
-  SpanDataAttributeValue operator()(unsigned int v)
-  {
-    return SpanDataAttributeValue(static_cast<uint64_t>(v));
-  }
-  SpanDataAttributeValue operator()(uint64_t v) { return SpanDataAttributeValue(v); }
-  SpanDataAttributeValue operator()(double v) { return SpanDataAttributeValue(v); }
-  SpanDataAttributeValue operator()(nostd::string_view v)
-  {
-    return SpanDataAttributeValue(std::string(v));
-  }
-  SpanDataAttributeValue operator()(nostd::span<const bool> v) { return convertSpan<bool>(v); }
-  SpanDataAttributeValue operator()(nostd::span<const int64_t> v)
-  {
-    return convertSpan<int64_t>(v);
-  }
-  SpanDataAttributeValue operator()(nostd::span<const unsigned int> v)
-  {
-    return convertSpan<uint64_t>(v);
-  }
-  SpanDataAttributeValue operator()(nostd::span<const uint64_t> v)
-  {
-    return convertSpan<uint64_t>(v);
-  }
-  SpanDataAttributeValue operator()(nostd::span<const double> v) { return convertSpan<double>(v); }
-  SpanDataAttributeValue operator()(nostd::span<const int> v) { return convertSpan<int64_t>(v); }
-  SpanDataAttributeValue operator()(nostd::span<const nostd::string_view> v)
-  {
-    return convertSpan<std::string>(v);
-  }
-
-  template <typename T, typename U = T>
-  SpanDataAttributeValue convertSpan(nostd::span<const U> vals)
-  {
-    std::vector<T> copy;
-    for (auto &val : vals)
-    {
-      copy.push_back(T(val));
-    }
-
-    return SpanDataAttributeValue(std::move(copy));
-  }
-};
-
-/**
- * Class for storing attributes.
- */
-class AttributeMap
-{
-public:
-  // Contruct empty attribute map
-  AttributeMap(){};
-
-  // Contruct attribute map and populate with attributes
-  AttributeMap(const trace_api::KeyValueIterable &attributes)
-  {
-    attributes.ForEachKeyValue(
-        [&](nostd::string_view key, opentelemetry::common::AttributeValue value) noexcept {
-          SetAttribute(key, value);
-          return true;
-        });
-  }
-
-  const std::unordered_map<std::string, SpanDataAttributeValue> &GetAttributes() const noexcept
-  {
-    return attributes_;
-  }
-
-  void SetAttribute(nostd::string_view key,
-                    const opentelemetry::common::AttributeValue &value) noexcept
-  {
-    attributes_[std::string(key)] = nostd::visit(converter_, value);
-  }
-
-private:
-  std::unordered_map<std::string, SpanDataAttributeValue> attributes_;
-  AttributeConverter converter_;
-};
-
 /**
  * Class for storing events in SpanData.
  */
@@ -283,7 +180,7 @@ public:
   {
     SpanDataLink link(attributes);
     links_.push_back(link);
-    // TODO: Add trace_id, span_id and trace_state when these are supported by SpanContext
+    // TODO: Populate trace_id, span_id and trace_state when these are supported by SpanContext
   }
 
   void SetStatus(trace_api::CanonicalCode code, nostd::string_view description) noexcept override
@@ -313,7 +210,6 @@ private:
   AttributeMap attribute_map_;
   std::vector<SpanDataEvent> events_;
   std::vector<SpanDataLink> links_;
-  AttributeConverter converter_;
 };
 }  // namespace trace
 }  // namespace sdk
